restructure wip
