+++ transcript_manager/CMakeLists.txt
@@ -31,6 +31,8 @@ endforeach()
 add_library(${PROJECT_NAME} SHARED
   src/transcript_manager_node.cpp
   src/transcript.cpp
+  src/transcript_operations.cpp
+  src/transcript_algorithms.cpp
 )
 ament_target_dependencies(${PROJECT_NAME}
   ${TRANSCRIPT_MANAGER_DEPENDENCIES}
